+++ station-ws/src/main/java/org/binas/station/domain/UserBalance.java
@@ -4,19 +4,19 @@ package org.binas.station.domain;
 
 public class UserBalance {
 
-	private double _value;
+	private int _value;
 	private int _tag;
 	
-	public UserBalance(double value, int tag) {
+	public UserBalance(int value, int tag) {
 		set_value(value);
 		set_tag(tag);
 	}
 
-	public double get_value() {
+	public int get_value() {
 		return _value;
 	}
 
-	public void set_value(double _value) {
+	public void set_value(int _value) {
 		this._value = _value;
 	}
